Add retweet functionnality

+++ src/services/tweets.ts
@@ -3,7 +3,12 @@ import { BaseUser } from '../types/User';
 import { supabaseClient } from './supabase';
 
 export function retrieveTweets() {
-  return supabaseClient.from('tweets').select('*, user:users(*), likes(*)');
+  return supabaseClient
+    .from('tweets')
+    .select(
+      '*, user:users(*), original:origin_tweet(*, user:users(*)), likes(*), retweets:tweets(*)'
+    )
+    .order('created_at', { ascending: false });
 }
 
 export function like(tweet: BaseTweet, user: BaseUser) {
@@ -35,6 +40,45 @@ export function like(tweet: BaseTweet, user: BaseUser) {
           .from('likes')
           .delete()
           .eq('tweet_id', tweet.id)
+          .eq('user_id', user.id)
+          .then(function () {
+            return null;
+          });
+      }
+    });
+}
+
+export function retweet(tweet: BaseTweet, user: BaseUser) {
+  return supabaseClient
+    .from('tweets')
+    .select('*')
+    .eq('origin_tweet', tweet.id)
+    .eq('user_id', user.id)
+    .then(function (resp) {
+      if (resp.data?.length === 0) {
+        return supabaseClient
+          .from('tweets')
+          .insert([
+            {
+              origin_tweet: tweet.id,
+              user_id: user.id,
+            },
+          ])
+          .select(
+            '*, user:users(*), original:origin_tweet(*, user:users(*)), likes(*), retweets:tweets(*)'
+          )
+          .then(function (resp) {
+            if (resp.data) {
+              return resp.data[0];
+            }
+
+            return null;
+          });
+      } else {
+        return supabaseClient
+          .from('tweets')
+          .delete()
+          .eq('origin_tweet', tweet.id)
           .eq('user_id', user.id)
           .then(function () {
             return null;
